toxicscanner/requirements.txt
flask
transformers
torch
pandas
python-dotenv
requests
